backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
mysqlclient>=2.1.1
python-dotenv
gunicorn
requests
numpy
qdrant-client
openai
python-decouple
pytest
rapidfuzz
drf-spectacular